Change input type from int to float for precision

+++ exercicios/ex008.py
@@ -1,2 +1,2 @@
-x = int(input('Digite um valor (em metros): '))
+x = float(input('Digite um valor (em metros): '))
 print('{} metro(s) equivale(m) a {} centímetro(s) e a {} milímetro(s).'.format(x, x*100, x*1000))
